+++ releaf-core/app/lib/releaf/core/responders/access_denied_responder.rb
@@ -1,6 +1,6 @@
-module Releaf::Responders
+module Releaf::Core::Responders
   class AccessDeniedResponder < ActionController::Responder
-    include Releaf::Responders::ErrorResponder
+    include Releaf::Core::Responders::ErrorResponder
 
     def status_code
       403