+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Renamed `Releaf::TemplateFieldTypeMapper` to `Releaf::Core::TemplateFieldTypeMapper`
 * Renamed `Releaf::AssetsResolver` to `Releaf::Core::AssetsResolver`
 * Renamed `Releaf::ErrorFormatter` to `Releaf::Core::ErrorFormatter`
+* Moved `Releaf::Responders` under `Releaf::Core` namespace
 
 ### 2015.08.04
 * refactored `@searchable_fields`. Now you should override `#searchable_fields`